refactoring; don't close OutputStream passed from the caller; better handle IOException during flushing
